Moving all common subscriber logic within observables to common classes
All common subscriber logic has been moved to
CompletableSubscriberWrapper. SingleSubscriberWrapper and
StreamSubscriberWrapper both extend from this to add custom logic for
onNext and onItem
